refactor(observability): downshift server tracing/correlation middleware to observability (P4a Batch 1b)
Move the request-scoped tracing and correlation middleware (and their internal
dependencies otel_bridge + request_logging) DOWN from aragora.server.middleware
to aragora.observability.middleware so foundation/infrastructure modules can
reach them without importing aragora.server. Transparent re-export shims at the
old paths keep importing and emit a DeprecationWarning (two-sided; ledgered in
library/shims.md).

- New package aragora/observability/middleware/ holds the moved modules;
  intra-cluster sibling imports use the new absolute paths (matching the
  original convention).
- aragora/server/middleware/{tracing,correlation,otel_bridge,request_logging}.py
  are deprecation shims that alias the relocated module via
  `sys.modules[__name__] = _target` after warning, so attribute access,
  from-imports, and existing monkeypatch-based tests all resolve to the single
  moved module while the old import path still emits a DeprecationWarning.
- aragora/server/middleware/__init__.py re-exports from the new location
  (server -> observability is a downward import; avoids self-warn).
- Foundation/infra importers repointed off the old paths:
  events.{handler_events,types}, queue.tracing, observability.{otel,trace_correlation}.
- events.{dispatcher,async_dispatcher} and observability.tracing keep importing
  via the shim for now: they carry pre-existing latent mypy debt unrelated to
  this move, so their repoint is deferred to the mypy-paydown phase.
- Domain/application/interface callers (knowledge, debate, nomic, server/*,
  logging_config) intentionally remain on the shim (handled by later batches).
- Tests: repointed the import-interception sites that simulate the tracing
  middleware (test_otel_bridge.py, test_handler_events.py,
  test_event_pipeline_e2e.py) to the new module path, since the relocated code
  imports its sibling via aragora.observability.middleware.tracing.

No import-contracts baseline change: none of the events/observability/queue ->
aragora.server edges fully clears this batch (each retains other-batch
contributors). make test-smoke passes (exit 0); targeted middleware/observability
tests green.
