+++ aragora/observability/trace_correlation.py
@@ -88,7 +88,7 @@ def get_trace_context() -> TraceContext:
         TraceContext with trace_id, span_id, and sampling decision
     """
     try:
-        from aragora.server.middleware.tracing import get_trace_id, get_span_id
+        from aragora.observability.middleware.tracing import get_trace_id, get_span_id
 
         trace_id = get_trace_id()
         span_id = get_span_id()
